refactor(terminal): extract terminal utilities to dedicated @ccusage/terminal package
This is a significant refactoring to improve the monorepo structure by extracting
reusable terminal utilities into a dedicated package.

## Changes made:

### New @ccusage/terminal package created
- Created new private package at packages/terminal
- Exports two modules: /table and /utils
- Preserves all in-source tests with vitest
- Logger is now passed as a constructor option (defaults to console.warn)

### Files moved:
- apps/ccusage/src/_terminal-utils.ts → packages/terminal/src/utils.ts
- apps/ccusage/src/_table.ts → packages/terminal/src/table.ts

### Dependencies reorganized:
- cli-table3 moved to @ccusage/terminal dependencies
- ansi-escapes and string-width remain in both packages as needed
- @oxc-project/runtime kept in dependencies for potential future use

### All imports updated across ccusage:
- Updated 12 files to import from @ccusage/terminal/table
- Updated 2 files to import from @ccusage/terminal/utils
- All tests passing (275 tests in ccusage, 30 tests in terminal)

### Configuration files added:
- ESLint config with svelte disabled
- TypeScript config extending root config
- Vitest config for in-source testing

## Testing:
- All existing tests preserved and passing
- Type checking successful
- ESLint formatting applied
