+++ apps/ccusage/src/commands/blocks.ts
@@ -14,7 +14,7 @@ import {
 
 } from '../_session-blocks.ts';
 import { sharedCommandConfig } from '../_shared-args.ts';
-import { formatCurrency, formatModelsDisplayMultiline, formatNumber, ResponsiveTable } from '../_table.ts';
+import { formatCurrency, formatModelsDisplayMultiline, formatNumber, ResponsiveTable } from '@ccusage/terminal/table';
 import { getTotalTokens } from '../_token-utils.ts';
 import { getClaudePaths, loadSessionBlockData } from '../data-loader.ts';
 import { log, logger } from '../logger.ts';
