+++ packages/terminal/eslint.config.js
@@ -1,0 +1,5 @@
+import { ryoppippi } from '@ryoppippi/eslint-config';
+
+export default ryoppippi({
+	svelte: false,
+});
