+++ packages/terminal/src/utils.ts
